Refactor route paths in App.jsx and AddEvents.jsx
File Organized
