+++ Admin-Front End/src/pages/ForgotPwd/FrameComponent.jsx
