+++ Admin-Front End/src/pages/ForgotPwd/Root1.jsx
